chore(storybook): extract the theme decorator into its own module
Move the ThemeBlock component and the withTheme decorator out of preview.tsx
into .storybook/withThemeDecorator.tsx, leaving preview.tsx with just Storybook
config. Addresses optional review feedback.
